Anim is now more fluid (4 images, was 2); It's now quite easy to choose anim at compile-time

+++ src/display/display.hpp
@@ -116,7 +116,7 @@ namespace bombherman
 		static SDL_Surface *gBack;
 		
 		static int gMapSize;
-		static Uint32 gSize;
+		static Uint16 gSize;
 		static SDL_Rect gZone;
 	};
 }
